DdeInitializeW, DdeGetLastError, etc. use UINT as status return type
(As a follow-up, DdeConnection::GetError should probably also be changed from
returning tools::Long to something like sal_uInt32.)

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/106581
Tested-by: Jenkins

+++ svl/source/svdde/ddecli.cxx
@@ -52,7 +52,7 @@ void ImpDeinitInstData()
 struct DdeImp
 {
     HCONV   hConv;
-    long    nStatus;
+    UINT    nStatus;
 };
 
 HDDEDATA CALLBACK DdeInternal::CliCallback( UINT nCode, UINT nCbType,
